+++ chpl/chpl-service/src/main/java/gov/healthit/chpl/domain/search/SearchViews.java
@@ -4,4 +4,10 @@ public class SearchViews {
 	public static class Default {
 		public static final String NAME = "DEFAULT";
 	}
+	public static class DecertifiedListings {
+		public static final String NAME = "DECERTIFIED LISTINGS";
+	}
+	public static class InactiveListings {
+		public static final String NAME = "INACTIVE LISTINGS";
+	}
 }